Assistant checkpoint: Create migrations to convert institution IDs to integers
Assistant generated file changes:
- lib/db/migrations/v14_convert_institution_ids_to_incremental.sql: Convert institution category and type IDs from UUID to incremental integers
- prisma/schema.prisma: Update institution category and type IDs to use integers, Update foreign key references to use integer IDs, Update institution profile foreign key reference
- lib/db/migrations/v13_seed_institution_categories_and_types.sql: Update seed data to work with integer IDs
- lib/db/migrations/v15_reseed_institution_data.sql: Re-seed institution data with integer IDs

---

User prompt:

The institution category id and institution type id should be incremental id and not UUID. Please share the migration sql file to update the same in db

+++ lib/db/migrations/v13_seed_institution_categories_and_types.sql
@@ -16,7 +16,7 @@ ON CONFLICT (slug) DO NOTHING;
 -- Insert institution types for Traditional Educational Institutions
 INSERT INTO institution_types (category_id, name, slug, description)
 SELECT 
-  ic.id,
+  ic.id::integer,
   type_data.name,
   type_data.slug,
   type_data.description
